src/deia/dashboard/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
watchdog>=6.0.0
pyyaml==6.0.1
